requirements.txt
selenium
webdriver_manager
mysql.connector